\name{plot.PointAttribute-methods}
\docType{methods}
\alias{plot.PointAttribute-methods}
\alias{plot.PointAttribute,SpatialPointsDataFrame,character,numeric,numeric-method}
\alias{plot.PointAttribute}
\title{Plotting attributes of \code{SpatialPointsDataFrame} objects}
\description{
S4 Method for plotting attributes of \code{SpatialPointsDataFrame} objects.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "SpatialPointsDataFrame", y = "character", dist = "numeric", cex = "numeric")}}{}
\item{\code{x}}{A \code{"SpatialPointsDataFrame"} object from where the coordinates of the attribute will be retrieved.}
\item{\code{y}}{A \code{"character"} with the name of the attribute.}
\item{\code{dist}}{A \code{"numeric"} with the distance from the coordinate to plot the attribute text.}
\item{\code{cex}}{A \code{"numeric"} with the relative size to plot the attribute text.}
}}
\keyword{methods}
\keyword{PointAttribute}
\keyword{plot.PointAttribute}
%\keyword{ ~~ other possible keyword(s) ~~ }
