\name{plot.PolyLineAttribute-methods}
\docType{methods}
\alias{plot.PolyLineAttribute-methods}
\alias{plot.PolyLineAttribute,SpatialPolygonsDataFrame,character,numeric,numeric-method}
\alias{plot.PolyLineAttribute,SpatialLinesDataFrame,character,numeric,numeric-method}
\alias{plot.PolyLineAttribute}
\title{Plot attributes of \code{Spatial-Lines,Polygons-DataFrame} objects.}
\description{
S4 Method for plotting attributes of \code{SpatialLinesDataFrame} and \code{SpatialPolygonsDataFrame} objects.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "SpatialPolygonsDataFrame", y = "character", dist = "numeric", cex = "numeric")}}{}
\item{\code{x}}{\code{"SpatialPointsDataFrame"} or \code{"SpatialPointsDataFrame"} object from where the coordinates of the attribute will be retrieved.}
\item{\code{y}}{\code{"character"} with the name of the attribute.}
\item{\code{dist}}{\code{"numeric"} with the distance from the coordinate to plot the attribute text.}
\item{\code{cex}}{\code{"numeric"} with the relative size to plot the attribute text.}
}}
\keyword{methods}
\keyword{PolyLineAttribute}
\keyword{plot.PolyLineAttribute}
%\keyword{ ~~ other possible keyword(s) ~~ }
